Create Week4_Lab3
For this assignment, you have been hired by Sunny Side Casino. They want to break down the amount of money received to coin increments. However, they feel that only those that can use the machine must have winnings greater than $100.00. If they do not enter that amount, let them know that they need to enter more than $100 or go to another machine.
With the conversion, that means the paper bills will have to be converted to dollar coins. In fact, you are going to convert everything to change. Once again, it is important to state that you will need the user to enter at least 100 dollars. In addition to dollar coins, we will need to break it down into half dollars, dimes, nickels, and pennies. There is a quarter shortage, and the substitution is that half dollars are being used.
We are going to start with dollars and decrement the change amount as we progress through the application. Pennies will be used last. Remember, the amount needs to be the same as entered.
Requirements: 
1. Create pseudocode. What steps are needed to make this a working program? (5 Points)
2. Create the Java coding. Create in your own style and words the code that is needed to generate this result. Make sure that you post your name and this assignment at the top of your comments. Make sure you leave comments when needed. (25 Points)
3. Display the result. Provide at least 2 screen shots of the working application. (5 Points)
4. Summarize. The client of the casino is inquiring what is the program going to do? Let them know in just a few sentences. (5 Points)
5. Will there be a need for maintenance or any updates in the future? (5 Points)
6. Is there a need to train anyone to run this program? Answer in a few sentences. (5 Points)